Refactor: XR 통신 - DTO 수정

+++ src/main/java/com/mtvs/sejong/playlog/dto/PlayLogRequestDTO.java
@@ -11,5 +11,5 @@ public class PlayLogRequestDTO {
 
     @JsonProperty("roomNumber") // JSON에서 대문자를 사용하여 역직렬화 문제 해결
     private int roomNumber;
-    private long gameId;
+    private int gameId;
 }
